requirements.txt
requests
python-dotenv
sqlalchemy
schedule